pyproject.toml
[tool.poetry]
name = "instachatico"
version = "0.1.0"
description = ""
authors = ["Mikhail Denisov <denisovmvtmb@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.117.1"}
uvicorn = {extras = ["standard"], version = "^0.36.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
python-dotenv = "^1.1.1"
alembic = "^1.16.5"
pydantic-settings = "^2.10.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
langchain = {extras = ["openai"], version = "^0.3.27"}
celery = "^5.5.3"
redis = "^6.4.0"
langchain-community = "^0.3.29"
aiohttp = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
